added and renamed printer files and helper
