+++ src/cricket.py
@@ -1,4 +1,4 @@
-import src.printer.fixtureprinter as fp
+import src.printer.fixtures as f
 import src.cricket_enums as ce
 from src.cricket_team import CricketTeam as ct
 from src.reader.playcricket import parse_play_cricket_data
@@ -7,6 +7,6 @@ from src.reader.googlecalendar import parse_google_calendar_data
 def main():
     list_of_play_cricket_fixtures = parse_play_cricket_data()
     list_of_google_calendar_fixtures = parse_google_calendar_data()
-    fp.print_google_calendar_csv(list_of_play_cricket_fixtures, ce.Ground.DP)
+    f.print_google_calendar_csv(list_of_play_cricket_fixtures, ce.Ground.DP)
 
 main()
